requirements.txt
Django==3.1.7
django_googledrive_storage==1.6.0
django_mysql==4.5.0
djangorestframework==3.14.0
pandas
PyJWT==2.1.0
